Added support for FakeSlots and SyncableInventory
"Fake", "Ghost" or "Phantom" slots are called that way because they
do not allow direct stack interaction as in vanilla, but behave differently
when clicked. This system is often used to set "filters" for items without
actually using up the item in the players hand.

Buildcraft does this by preventing the Slots from being manipulated as usual,
but instead sends a packet to the server containing the details of the click
and the hold item.

This pull requests replicates this mechanism and exposes the system via
FakeSlot, FakeGuiContainer and IFakeSlotListener.
It also includes a straight forward SyncableInventory implementation extending
GenericInventory.

A usage example for both can be seen in an upcoming pull request to OP-Addons.